Reports: Add a YAML option to generate a report as part of a CBT run
Currently generating a performance report is a separate, manual step after a CBT benchmark runhas completed. This adds an option to the CBT configuration YAML file to automatically generate a performance report at the end of a benchmark run. This also requires a 'report' section in the YAML file to configure the options for the report.

Signed-off-by: Chris Harris <harriscr@uk.ibm.com>